+++ casual/casual-event-server/src/main/java/se/laz/casual/event/server/ServerInitialization.java
@@ -1,0 +1,9 @@
+package se.laz.casual.event.server;
+
+import io.netty.channel.Channel;
+import io.netty.channel.group.ChannelGroup;
+
+public interface ServerInitialization
+{
+    Channel init(EventServerConnectionInformation connectionInformation, ChannelGroup connectedClients);
+}
